requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
charset-normalizer==3.2.0
code-stego==0.1.0
discord.py==2.3.2
Django==4.0
frozenlist==1.4.0
gunicorn==21.2.0
idna==3.4
multidict==6.0.4
numpy==1.25.2
packaging==23.1
Pillow==10.0.0
Pygments==2.16.1
python-dotenv==1.0.0
sqlparse==0.4.4
tzdata==2023.3
yarl==1.9.2
